+++ infrastructure/modules/monitoring/keda/outputs.tf
@@ -1,5 +1,0 @@
-output "keda" {
-  value = {
-    chart_name = helm_release.keda.name
-  }
-}