+++ infrastructure/modules/needed-storage/outputs.tf
@@ -1,9 +1,0 @@
-# Needed resource for each storage
-output "needed_storage" {
-  value = local.needed_storage
-}
-
-# List of resources to deploy
-output "list_storage" {
-  value = local.list_storage
-}