+++ infrastructure/modules/onpremise-storage/minio/locals.tf
@@ -1,5 +1,0 @@
-locals {
-  node_selector_keys   = keys(var.minio.node_selector)
-  node_selector_values = values(var.minio.node_selector)
-  port                 = 9000
-}