+++ infrastructure/quick-deploy/localhost/all/storage.tf
@@ -1,6 +1,6 @@
 # ActiveMQ
 module "activemq" {
-  source      = "../../../modules/onpremise-storage/activemq"
+  source      = "./generated/infra-modules/storage/onpremise/activemq"
   namespace   = local.namespace
   working_dir = "${path.root}/../../.."
   activemq = {
@@ -13,7 +13,7 @@ module "activemq" {
 
 # MongoDB
 module "mongodb" {
-  source      = "../../../modules/onpremise-storage/mongodb"
+  source      = "./generated/infra-modules/storage/onpremise/mongodb"
   namespace   = local.namespace
   working_dir = "${path.root}/../../.."
   mongodb = {
@@ -29,7 +29,7 @@ module "mongodb" {
 # Redis
 module "redis" {
   count       = var.redis != null ? 1 : 0
-  source      = "../../../modules/onpremise-storage/redis"
+  source      = "./generated/infra-modules/storage/onpremise/redis"
   namespace   = local.namespace
   working_dir = "${path.root}/../../.."
   redis = {
@@ -44,7 +44,7 @@ module "redis" {
 # minio
 module "minio" {
   count     = var.minio != null ? 1 : 0
-  source    = "../../../modules/onpremise-storage/minio"
+  source    = "./generated/infra-modules/storage/onpremise/minio"
   namespace = local.namespace
   minio = {
     image              = var.minio.image_name
@@ -56,17 +56,28 @@ module "minio" {
   }
 }
 
+# minio for file storage
+module "minio_s3_fs" {
+  count     = var.minio_s3_fs != null ? 1 : 0
+  source    = "./generated/infra-modules/storage/onpremise/minio"
+  namespace = local.namespace
+  minio = {
+    image              = local.minio_s3_fs_image
+    tag                = try(coalesce(var.minio.image_tag), local.default_tags[var.minio_s3_fs.image_name])
+    image_pull_secrets = local.minio_s3_fs_image_pull_secrets
+    host               = local.minio_s3_fs_host
+    bucket_name        = local.minio_s3_fs_bucket_name
+    node_selector      = local.minio_s3_fs_node_selector
+  }
+}
+
 # Shared storage
 resource "kubernetes_secret" "shared_storage" {
   metadata {
     name      = "shared-storage"
     namespace = local.namespace
   }
-  data = {
-    host_path         = abspath("data")
-    file_storage_type = "HostPath"
-    file_server_ip    = ""
-  }
+  data = local.shared_storage
 }
 
 resource "kubernetes_secret" "deployed_object_storage" {