+++ infrastructure/utils/scripts/installation/prerequisites/install-make.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# This script is used to install make.
+
+sudo apt install -y make