+++ infrastructure/utils/scripts/installation/prerequisites/install-python3.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# This script is used to install python3.
+
+sudo apt install -y python3
+
+# Link python3 to python
+sudo ln -s /usr/bin/python3 /usr/bin/python || true